Add build script for vim-plug
Now it makes sense:

Upstream includes the build in the repo, but since this causes a ton of
merge conflicts, I prefer not to track build files in git. However in
this case, build needs to be done in the post-update hook of vim-plug.

Another way would be to build manually.